feat(preset): new PresetManager
